requirements.txt
webdriver-manager==4.0.2
selenium==4.38.0
pytest==8.4.0
pytest-html==4.1.0
pytest-xdist==3.6.0
allure-pytest==2.14.0
python-dotenv==1.1.0
requests==2.31.0
Faker==22.1.0
